backend/requirements.txt
sentence-transformers
faiss-cpu
pandas
numpy
pillow
pymupdf
pdf2image
pytesseract
Flask
Flask-Cors
